requirements.txt
kivy==2.3.0
kivymd==1.1.1
yfinance==0.2.36
pandas==2.1.4
numpy==1.26.2
plyer==2.1.0
pillow==10.1.0
requests==2.31.0